Add files via upload
This is our work on Airtel SheCodes Hackathon organised between 23-25th April 2024. 

The main objective of our problem statement was- Airtel Agents with necessary skills must reach the customers in shortest time possible.
To solve this problem we prepared a database schema. The database contained information about the customer, agents, problem type, and so on. 
Along with this , we employed A* algorithm to find the shortest path from agent to the customer's location.